todo-web/ai-service/pyproject.toml
[project]
name = "todo-ai-service"
version = "1.0.0"
description = "AI chatbot service with OpenAI Agents SDK and MCP - Phase III"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai-agents>=0.0.8",
    "mcp>=1.0.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "sse-starlette>=2.2.0",
]

[project.scripts]
serve = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8001"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py312"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
]
